requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.36
bcrypt==3.2.2
pymysql==1.1.2
python-dotenv==1.0.0
slowapi==0.1.9

